+++ src/DataStructures/types.h
@@ -34,10 +34,10 @@ class ImageMeasurement
 struct FRAMEINFO
 {
     EIGEN_MAKE_ALIGNED_OPERATOR_NEW
-    Eigen::Matrix3f R_k_2_c;//R_k^(k+1)
-    Eigen::Vector3f T_k_2_c;//T_k^(k+1)
+    Eigen::Matrix3d R_k_2_c;//R_k^(k+1)
+    Eigen::Vector3d T_k_2_c;//T_k^(k+1)
     //Matrix<float, 6, 6> lastestATA ;
-    Eigen::MatrixXf lastestATA ;
+    Eigen::MatrixXd lastestATA ;
     bool keyFrameFlag ;
     bool trust ;
     ros::Time t ;
@@ -46,9 +46,9 @@ struct FRAMEINFO
 struct CAMERALINK
 {
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
-  Eigen::Matrix3f R_bi_2_bj;//rotation from current body to link body
-  Eigen::Vector3f T_bi_2_bj;//translation from current body to link body
-  Eigen::Matrix<float, 6, 6> P_inv;
+  Eigen::Matrix3d R_bi_2_bj;//rotation from current body to link body
+  Eigen::Vector3d T_bi_2_bj;//translation from current body to link body
+  Eigen::Matrix<double, 6, 6> P_inv;
         //bool T_trust ;
         //MatrixXd P_inv;
 };
